Очень сонный коммит

+++ src/main/java/com/meow/bebrablender/render_engine/ModelNormalizer.java
@@ -1,6 +1,6 @@
 package com.meow.bebrablender.render_engine;
 
-import com.meow.bebrablender.math.vectors.Vector3f;
+import com.meow.bebrablender.math.vectors.Vector3d;
 import com.meow.bebrablender.model.Polygon;
 import com.meow.bebrablender.triangulation.TriangulationUtils;
 import com.meow.bebrablender.normals.NormalUtils;
@@ -14,9 +14,9 @@ public class ModelNormalizer {
      * @param normals  all model normals list.
      */
     public static void triangulateAndRecalculateModelNormals(
-            List<Vector3f> vertices,
+            List<Vector3d> vertices,
             List<Polygon> polygons,
-            List<Vector3f> normals) {
+            List<Vector3d> normals) {
         TriangulationUtils.triangulateModelPolygons(polygons);
         NormalUtils.recalculateModelNormals(vertices, normals, polygons);
     }
